refactor: extract `resolveUpgradeTargetVersion`
